fix: resolve 25 CodeQL scanning alerts in V2 code
- Add Severity comparison operators (eliminates 4 enum-undefined-behavior)
- Break path temporary chains into named variables (eliminates 15 use-after-expired-lifetime)
- Tighten bounds checks in exploit/security checks (eliminates 5 icc-buffer-overflow)
- Simplify enum comparisons (Command, icColorSpaceSignature)

+++ iccanalyzer-lite/icctest/tools/icctestParity.cpp
@@ -244,7 +244,7 @@ std::optional<Severity> maxFindingSeverity(const CheckResult& result) {
     }
     Severity maxLevel = result.findings.front().level;
     for (const auto& f : result.findings) {
-        if (static_cast<uint8_t>(f.level) > static_cast<uint8_t>(maxLevel)) {
+        if (f.level > maxLevel) {
             maxLevel = f.level;
         }
     }
